backend/requirements.txt
fastapi
python-multipart
uvicorn
msoffcrypto-tool
aiofiles
python-magic
cryptography
boto3
python-dotenv